requirements.txt
torch==2.5.1
transformers==4.47.1
vllm==0.6.6.post1
pyyaml==6.0.2
tqdm==4.67.1
accelerate==1.0.1
deepspeed==0.14.4
zhipuai==2.1.5.20250106